+++ packages/module/src/vite.ts
@@ -1,0 +1,13 @@
+export default function (reg = /\.module\.ts$/) {
+  return {
+    name: 'phecda-module-hmr',
+    enforce: 'pre',
+    apply: 'serve',
+    transform(code: string, id: string) {
+      if (reg.test(id)) {
+        return code += `\nif (import.meta.hot)
+    import.meta.hot.accept(__PHECDA_MODULE_UPDATE__)`
+      }
+    },
+  }
+}