+++ packages/server/src/core.ts
@@ -31,6 +31,7 @@ async function buildNestModule(Module: Construct, map: Map<string, InstanceType<
   const paramtypes = getParamtypes(Module) as Construct[]
   let instance: InstanceType<Construct>
   const tag = Module.prototype?.__TAG__ || Module.name
+
   if (map.has(tag)) {
     instance = map.get(tag)
     if (!instance)
@@ -60,28 +61,35 @@ function getMetaFromInstance(instance: Phecda, name: string, tag: string) {
   const vars = getExposeKey(instance).filter(item => item !== '__CLASS')
   const baseState = (getState(instance, '__CLASS') || {}) as P.Meta
   initState(baseState)
+
   return vars.map((i) => {
+    const meta = {} as P.Meta
     const state = (getState(instance, i) || {}) as P.Meta
-    if (baseState.route && state.route)
-      state.route.route = baseState.route.route + state.route.route
-    state.name = name
-    state.tag = tag
-    state.method = i
+    initState(state)
+    if (state.route) {
+      meta.route = {
+        route: (baseState.route?.route || '') + (state.route.route),
+        type: state.route.type,
+      }
+    }
+
+    meta.name = name
+    meta.tag = tag
+    meta.method = i as string
     const params = [] as any[]
     for (const i of state.params || []) {
       params.unshift(i)
       if (i.index === 0)
         break
     }
-    state.params = params
-    initState(state)
-    state.define = { ...baseState.define, ...state.define }
-    state.header = { ...baseState.header, ...state.header }
-    state.middlewares = [...new Set([...baseState.middlewares, ...state.middlewares])]
-    state.guards = [...new Set([...baseState.guards, ...state.guards])]
-    state.interceptors = [...new Set([...baseState.interceptors, ...state.interceptors])]
+    meta.params = params
+    meta.define = { ...baseState.define, ...state.define }
+    meta.header = { ...baseState.header, ...state.header }
+    meta.middlewares = [...new Set([...baseState.middlewares, ...state.middlewares])]
+    meta.guards = [...new Set([...baseState.guards, ...state.guards])]
+    meta.interceptors = [...new Set([...baseState.interceptors, ...state.interceptors])]
 
-    return new Meta(state as unknown as P.Meta, getHandler(instance, i), getParamtypes(instance, i) || [])
+    return new Meta(meta as unknown as P.Meta, getHandler(instance, i), getParamtypes(instance, i as string) || [])
   })
 }