+++ packages/server/src/decorators/route.ts
@@ -1,25 +1,19 @@
-import { mergeState, setModalVar } from 'phecda-core'
+import { setModelVar, setState } from 'phecda-core'
 
 export function Route(route: string, type?: string): any {
   return (target: any, key?: PropertyKey) => {
-    if (key) {
-      setModalVar(target, key)
-      mergeState(target, key, {
-        route: {
-          route,
-          type,
-        },
-      })
+    if (!key)
+      key = '__CLASS'
+    target = key === '__CLASS' ? target.prototype : target
+
+    setModelVar(target, key)
+
+    const state = target._namespace.__STATE_NAMESPACE__.get(key) || {}
+    state.route = {
+      route,
+      type,
     }
-    else {
-      setModalVar(target.prototype, '__CLASS')
-      mergeState(target.prototype, '__CLASS', {
-        route: {
-          route,
-          type,
-        },
-      })
-    }
+    setState(target, key, state)
   }
 }
 
@@ -44,51 +38,48 @@ export function Controller(route: string) {
 
 export function Guard(...guards: string[]): any {
   return (target: any, key?: PropertyKey) => {
-    if (key) {
-      setModalVar(target, key)
-      mergeState(target, key, {
-        guards: [...guards],
-      })
-    }
-    else {
-      setModalVar(target.prototype, '__CLASS')
-      mergeState(target.prototype, '__CLASS', {
-        guards: [...guards],
-      })
-    }
+    if (!key)
+      key = '__CLASS'
+    target = key === '__CLASS' ? target.prototype : target
+
+    setModelVar(target, key)
+
+    const state = target._namespace.__STATE_NAMESPACE__.get(key) || {}
+    if (!state.guards)
+      state.guards = []
+    state.guards.push(...guards)
+    setState(target, key, state)
   }
 }
 
 export function Middle(...middlewares: string[]): any {
   return (target: any, key?: PropertyKey) => {
-    if (key) {
-      setModalVar(target, key)
-      mergeState(target, key, {
-        middlewares: [...middlewares],
-      })
-    }
-    else {
-      setModalVar(target.prototype, '__CLASS')
-      mergeState(target.prototype, '__CLASS', {
-        middlewares: [...middlewares],
-      })
-    }
+    if (!key)
+      key = '__CLASS'
+    target = key === '__CLASS' ? target.prototype : target
+
+    setModelVar(target, key)
+
+    const state = target._namespace.__STATE_NAMESPACE__.get(key) || {}
+    if (!state.middlewares)
+      state.middlewares = []
+    state.middlewares.push(...middlewares)
+    setState(target, key, state)
   }
 }
 
 export function Interceptor(...interceptors: string[]): any {
   return (target: any, key?: PropertyKey) => {
-    if (key) {
-      setModalVar(target, key)
-      mergeState(target, key, {
-        interceptors: [...interceptors],
-      })
-    }
-    else {
-      setModalVar(target.prototype, '__CLASS')
-      mergeState(target.prototype, '__CLASS', {
-        interceptors: [...interceptors],
-      })
-    }
+    if (!key)
+      key = '__CLASS'
+    target = key === '__CLASS' ? target.prototype : target
+
+    setModelVar(target, key)
+
+    const state = target._namespace.__STATE_NAMESPACE__.get(key) || {}
+    if (!state.interceptors)
+      state.interceptors = []
+    state.interceptors.push(...interceptors)
+    setState(target, key, state)
   }
 }