+++ packages/vue/src/components/createTable.ts
@@ -4,34 +4,48 @@ import type { Component, DefineComponent } from 'vue'
 
 export function createTable<P extends { $props: any }>(
   compSet: Record<string, Component> | any,
-  table: Component<P>,
-  tableColumn: Component | false,
-  data?: any,
+  options: {
+    table: Component<P>
+    column?: Component | false
+    data?: any
+    columnSlot?: string
+    cellSlot?: string
+  },
+
 ): DefineComponent<{
   config: Object
 } & P['$props']> {
+  const { table, column: tableColumn, data, cellSlot = 'default', columnSlot = 'default' } = options
   const TableColumn = defineComponent({
     name: 'PhecdaTableColumn',
-    props: {
-      tableColumn: {
-        type: Object, required: true,
-      },
 
-    },
-    setup(props) {
-      const compName = props.tableColumn._component
-      return () =>
-        h(
+    setup(_props, { attrs }) {
+      const compName = attrs._component as string
+      const childrens: any = attrs._children && {
+        default: () => (attrs._children as any[]).map((child: any) => h(tableColumn as any, child)),
+      }
+
+      if (tableColumn) {
+        return () => h(
           tableColumn as any,
-          {
-            ...props.tableColumn,
-          },
-          {
-            default: (scope: any) => {
-              return compName ? h(compSet[compName], { scope, data, ...props.tableColumn._props || {} }) : null
-            },
-          },
+          attrs,
+          compName
+            ? {
+                [`${cellSlot}`]: (scope: any) => {
+                  const childrenProps = typeof attrs._props === 'function' ? attrs._props?.(scope) : attrs._props
+                  return h(compSet[compName], { data, ...(childrenProps || {}) }, childrenProps?._slots)
+                },
+              }
+            : childrens,
         )
+      }
+      else {
+        return (scope: any) => {
+          const childrenProps = typeof attrs._props === 'function' ? attrs._props?.(scope) : attrs._props
+
+          return compName ? h(compSet[compName], { scope, data, ...(childrenProps || {}) }, childrenProps?._slots) : childrens
+        }
+      }
     }
     ,
   })
@@ -51,11 +65,10 @@ export function createTable<P extends { $props: any }>(
       })
       return () => {
         return h(table as any, Object.assign({ ref: dom }, ctx.attrs), {
-          default: () =>
+          [`${columnSlot}`]: (scope: any) =>
             props.config.map((item: any) => {
-              return h(TableColumn as any, {
-                tableColumn: item,
-              })
+              item = typeof item === 'function' ? item(scope) : item
+              return h(TableColumn, item)
             }),
         })
       }