+++ packages/rabbitmq/CHANGELOG.md
@@ -1,5 +1,12 @@
 # phecda-rabbitmq
 
+## 1.0.2-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [78cb57a]
+  - phecda-server@3.0.0-alpha.0
+
 ## 1.0.1
 
 ### Patch Changes